pyproject.toml
[project]
name = "github-blog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "feedgen>=1.0.0",
    "jinja2>=3.1.4",
    "marko>=2.1.2",
    "pygithub>=2.4.0",
]
